pyproject.toml
[project]
name = "vllm-local-qwen-langsmith"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langgraph>=1.1.6",
    "langgraph-cli[inmem]>=0.4.21",
    "langsmith>=0.7.30",
    "openai>=2.31.0",
    "python-dotenv>=1.2.2",
    "vllm>=0.19.0",
]
